Move AuthInterceptor to core.network package
